+++ example/src/main/java/example/ExampleCsvReaderWithCustomSimpleMapper.java
@@ -1,7 +1,6 @@
 package example;
 
 import de.siegmar.fastcsv.reader.CsvCallbackHandler;
-import de.siegmar.fastcsv.reader.CsvCallbackHandlers;
 import de.siegmar.fastcsv.reader.CsvReader;
 
 /**
@@ -12,7 +11,7 @@ public class ExampleCsvReaderWithCustomSimpleMapper {
     private static final String DATA = "john smith,56\njane doe,38";
 
     public static void main(final String[] args) {
-        final CsvCallbackHandler<Person> mapper = CsvCallbackHandlers.forSimpleMapper(
+        final CsvCallbackHandler<Person> mapper = CsvCallbackHandler.forSimpleMapper(
             fields -> new Person(fields[0], Integer.parseInt(fields[1]))
         );