+++ example/src/main/java/example/ExampleCsvReaderWithFieldModifier.java
@@ -3,6 +3,7 @@ package example;
 import java.util.Locale;
 
 import de.siegmar.fastcsv.reader.CsvReader;
+import de.siegmar.fastcsv.reader.CsvRecordHandler;
 import de.siegmar.fastcsv.reader.FieldModifier;
 import de.siegmar.fastcsv.reader.FieldModifiers;
 
@@ -15,40 +16,30 @@ public class ExampleCsvReaderWithFieldModifier {
 
     public static void main(final String[] args) {
         System.out.println("Trim fields:");
-        builderWithTrim().ofCsvRecord(DATA)
+        CsvReader.builder().build(new CsvRecordHandler(FieldModifiers.TRIM), DATA)
             .forEach(System.out::println);
 
         System.out.println("Trim and lowercase fields:");
-        builderWithTrimAndLowerCase().ofCsvRecord(DATA)
+        CsvReader.builder().build(new CsvRecordHandler(combinedModifier()), DATA)
             .forEach(System.out::println);
 
         System.out.println("Trim and lowercase fields of first record (by using a custom modifier):");
-        builderWithCustomModifier().ofCsvRecord(DATA)
+        CsvReader.builder().build(new CsvRecordHandler(customModifier()), DATA)
             .forEach(System.out::println);
     }
 
-    private static CsvReader.CsvReaderBuilder builderWithTrim() {
-        return CsvReader.builder().fieldModifier(FieldModifiers.TRIM);
+    private static FieldModifier combinedModifier() {
+        return FieldModifiers.TRIM.andThen(FieldModifiers.lower(Locale.ENGLISH));
     }
 
-    private static CsvReader.CsvReaderBuilder builderWithTrimAndLowerCase() {
-        final FieldModifier modifier = FieldModifiers.TRIM.andThen(FieldModifiers.lower(Locale.ENGLISH));
-        return CsvReader.builder().fieldModifier(modifier);
-    }
-
-    private static CsvReader.CsvReaderBuilder builderWithCustomModifier() {
-        final FieldModifier modifier = new FieldModifier() {
+    private static FieldModifier customModifier() {
+        return new FieldModifier() {
             @Override
             public String modify(final long startingLineNumber, final int fieldIdx, final boolean quoted,
                                  final String field) {
-                if (startingLineNumber == 1) {
-                    return field.trim().toLowerCase(Locale.ENGLISH);
-                }
-                return field;
+                return startingLineNumber == 1 ? field.trim().toLowerCase(Locale.ENGLISH) : field;
             }
         };
-
-        return CsvReader.builder().fieldModifier(modifier);
     }
 
 }