+++ example/src/main/java/example/ExampleNamedCsvReaderWithCustomHeader.java
@@ -1,9 +1,9 @@
 package example;
 
 import de.siegmar.fastcsv.reader.CsvCallbackHandler;
-import de.siegmar.fastcsv.reader.CsvCallbackHandlers;
 import de.siegmar.fastcsv.reader.CsvReader;
 import de.siegmar.fastcsv.reader.NamedCsvRecord;
+import de.siegmar.fastcsv.reader.NamedCsvRecordHandler;
 
 /**
  * Example for reading CSV data with a custom header.
@@ -14,7 +14,7 @@ public class ExampleNamedCsvReaderWithCustomHeader {
 
     public static void main(final String[] args) {
         final CsvCallbackHandler<NamedCsvRecord> callbackHandler =
-            CsvCallbackHandlers.ofNamedCsvRecord("header1", "header2");
+            new NamedCsvRecordHandler("header1", "header2");
 
         CsvReader.builder().build(callbackHandler, DATA)
             .forEach(rec -> System.out.println(rec.getField("header2")));