+++ lib/src/intTest/java/blackbox/reader/CsvReaderBuilderTest.java
@@ -78,7 +78,7 @@ class CsvReaderBuilderTest {
         assertThat(crb).asString()
             .isEqualTo("CsvReaderBuilder[fieldSeparator=,, quoteCharacter=\", "
                 + "commentStrategy=NONE, commentCharacter=#, skipEmptyLines=true, "
-                + "ignoreDifferentFieldCount=true, fieldModifier=null]");
+                + "ignoreDifferentFieldCount=true]");
     }
 
     @Test