+++ lib/src/main/java/de/siegmar/fastcsv/reader/AbstractCsvCallbackHandler.java
@@ -1,5 +1,7 @@
 package de.siegmar.fastcsv.reader;
 
+import java.util.Objects;
+
 import de.siegmar.fastcsv.util.Limits;
 
 /**
@@ -10,6 +12,11 @@ import de.siegmar.fastcsv.util.Limits;
 public abstract class AbstractCsvCallbackHandler<T> implements CsvCallbackHandler<T> {
 
     private static final int INITIAL_FIELDS_SIZE = 32;
+
+    /**
+     * The field modifier.
+     */
+    protected final FieldModifier fieldModifier;
 
     /**
      * The starting line number of the current record.
@@ -42,10 +49,22 @@ public abstract class AbstractCsvCallbackHandler<T> implements CsvCallbackHandle
      * Constructs a new instance with an initial fields array of size {@value #INITIAL_FIELDS_SIZE}.
      */
     protected AbstractCsvCallbackHandler() {
-        this(INITIAL_FIELDS_SIZE);
+        this(FieldModifiers.NOP, INITIAL_FIELDS_SIZE);
     }
 
-    AbstractCsvCallbackHandler(final int len) {
+    /**
+     * Constructs a new instance with the given field modifier and initial fields array of size
+     * {@value #INITIAL_FIELDS_SIZE}.
+     *
+     * @param fieldModifier the field modifier, must not be {@code null}
+     * @throws NullPointerException if {@code null} is passed
+     */
+    protected AbstractCsvCallbackHandler(final FieldModifier fieldModifier) {
+        this(fieldModifier, INITIAL_FIELDS_SIZE);
+    }
+
+    AbstractCsvCallbackHandler(final FieldModifier fieldModifier, final int len) {
+        this.fieldModifier = Objects.requireNonNull(fieldModifier, "fieldModifier must not be null");
         fields = new String[len];
     }
 
@@ -64,24 +83,59 @@ public abstract class AbstractCsvCallbackHandler<T> implements CsvCallbackHandle
 
     /**
      * {@inheritDoc}
-     * Adds the given value to the internal fields array.
-     * <p>
-     * Extends the array if necessary and keeps track of the total record size and fields count.
+     * Passes the materialized ({@link #materializeField(char[], int, int)}) and
+     * modified ({@link #modifyField(String, boolean)}) value to {@link #addField(String, boolean)}.
      *
      * @throws CsvParseException if the addition exceeds the limit of record size or maximum fields count.
      */
     @Override
-    public void addField(final String value, final boolean quoted) {
-        recordSize += value.length();
-        if (recordSize > Limits.MAX_RECORD_SIZE) {
+    public void addField(final char[] buf, final int offset, final int len, final boolean quoted) {
+        if (recordSize + len > Limits.MAX_RECORD_SIZE) {
             throw new CsvParseException(maxRecordSizeExceededMessage(startingLineNumber));
         }
+        addField(modifyField(materializeField(buf, offset, len), quoted), quoted);
+    }
+
+    /**
+     * Adds the given value to the internal fields array.
+     * <p>
+     * Extends the array if necessary and keeps track of the total record size and fields count.
+     *
+     * @param value  the field value
+     * @param quoted {@code true} if the field was quoted
+     * @throws CsvParseException if the addition exceeds the limit of record size or maximum fields count.
+     */
+    protected void addField(final String value, final boolean quoted) {
+        recordSize += value.length();
 
         if (idx == fields.length) {
             extendCapacity();
         }
 
         fields[idx++] = value;
+    }
+
+    /**
+     * Modifies field value.
+     *
+     * @param value  the field value
+     * @param quoted {@code true} if the field was quoted
+     * @return the modified field value
+     */
+    protected String modifyField(final String value, final boolean quoted) {
+        return fieldModifier.modify(startingLineNumber, idx, quoted, value);
+    }
+
+    /**
+     * Materializes field from the given buffer.
+     *
+     * @param buf    the internal buffer that contains the field value (among other data)
+     * @param offset the offset of the field value in the buffer
+     * @param len    the length of the field value
+     * @return the materialized field value
+     */
+    protected String materializeField(final char[] buf, final int offset, final int len) {
+        return new String(buf, offset, len);
     }
 
     private static String maxRecordSizeExceededMessage(final long line) {
@@ -91,26 +145,60 @@ public abstract class AbstractCsvCallbackHandler<T> implements CsvCallbackHandle
 
     /**
      * {@inheritDoc}
-     * Sets the given value as the only field in the internal fields array.
-     * <p>
-     * Keeps track of the total record size.
+     * Passes the materialized ({@link #materializeComment(char[], int, int)}) and
+     * modified ({@link #modifyComment(String)}) value to {@link #setComment(String)}.
      *
      * @throws CsvParseException if the addition exceeds the limit of record size.
      */
     @Override
-    public void setComment(final String value) {
-        recordSize += value.length();
-        if (recordSize > Limits.MAX_RECORD_SIZE) {
-            throw new CsvParseException(maxRecordSizeExceededMessage(startingLineNumber));
-        }
-
+    public void setComment(final char[] buf, final int offset, final int len) {
         if (idx != 0) {
             // Can't happen with the current implementation of CsvParser
             throw new IllegalStateException("Comment must be the first and only field in a record");
         }
 
+        if (recordSize + len > Limits.MAX_RECORD_SIZE) {
+            // Can't happen with the current implementation of CsvParser
+            throw new CsvParseException(maxRecordSizeExceededMessage(startingLineNumber));
+        }
+
+        setComment(modifyComment(materializeComment(buf, offset, len)));
+    }
+
+    /**
+     * Sets the given value as the only field in the internal fields array.
+     * <p>
+     * Keeps track of the total record size.
+     *
+     * @param value the comment value
+     * @throws CsvParseException if the addition exceeds the limit of record size.
+     */
+    protected void setComment(final String value) {
+        recordSize += value.length();
         comment = true;
         fields[idx++] = value;
+    }
+
+    /**
+     * Modifies comment value.
+     *
+     * @param field the comment value
+     * @return the modified comment value
+     */
+    protected String modifyComment(final String field) {
+        return fieldModifier.modifyComment(startingLineNumber, field);
+    }
+
+    /**
+     * Materializes comment from the given buffer.
+     *
+     * @param buf    the internal buffer that contains the comment value (among other data)
+     * @param offset the offset of the field value in the buffer
+     * @param len    the length of the field value
+     * @return the materialized field value
+     */
+    protected String materializeComment(final char[] buf, final int offset, final int len) {
+        return new String(buf, offset, len);
     }
 
     private void extendCapacity() {