+++ lib/src/main/java/de/siegmar/fastcsv/reader/CsvRecordHandler.java
@@ -1,6 +1,26 @@
 package de.siegmar.fastcsv.reader;
 
-final class CsvRecordHandler extends AbstractCsvCallbackHandler<CsvRecord> {
+/**
+ * A {@link CsvCallbackHandler} implementation that returns a {@link CsvRecord} for each record.
+ */
+public final class CsvRecordHandler extends AbstractCsvCallbackHandler<CsvRecord> {
+
+    /**
+     * Constructs a new {@code CsvRecordHandler}.
+     */
+    public CsvRecordHandler() {
+        super();
+    }
+
+    /**
+     * Constructs a new {@code CsvRecordHandler} with the given field modifier.
+     *
+     * @param fieldModifier the field modifier, must not be {@code null}
+     * @throws NullPointerException if {@code null} is passed
+     */
+    public CsvRecordHandler(final FieldModifier fieldModifier) {
+        super(fieldModifier);
+    }
 
     @Override
     protected CsvRecord buildRecord(final String[] fields) {