+++ lib/src/main/java/de/siegmar/fastcsv/reader/FieldModifier.java
@@ -1,12 +1,14 @@
 package de.siegmar.fastcsv.reader;
 
 /**
- * Implementations of this class are used to modify fields before they are passed to the {@link CsvCallbackHandler}
- * and finally stored in a record.
+ * Implementations of this class are used within {@link CsvCallbackHandler} implementations to modify the fields of
+ * a CSV record before storing them in the resulting object.
+ * <p>
+ * Applying field modifiers might affect the behavior of skipping empty lines – see
+ * {@link CsvReader.CsvReaderBuilder#skipEmptyLines(boolean)}.
  *
  * @see FieldModifiers
  * @see SimpleFieldModifier
- * @see CsvReader.CsvReaderBuilder#fieldModifier(FieldModifier)
  */
 public interface FieldModifier {