+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `IndexedCsvReader` for random access to CSV files
 - `FieldModifier` for modifying fields while reading CSV files
 - Allow custom quote strategies for CsvWriter
-- `CsvCallbackHandlers` for more flexible usage of CsvReader
+- `CsvCallbackHandler` for more flexible usage of CsvReader
 - Support for optional BOM header when reading CSV files
 - Method `NamedCsvRecord.findField` for optional field access
 - Allow READ comment strategy for CSV data with a header
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated from Java 8 to Java 11
 - Updated naming (rows/lines -> records, columns -> fields, differentiate between lines and records)
-- `NamedCsvReader` replaced by `CsvReader` with `CsvCallbackHandlers`
+- `NamedCsvReader` replaced by `CsvReader.builder().ofNamedCsvRecord()`
 - `build` methods in `CsvReaderBuilder` with callback handlers and `ofCsvRecord` / `ofNamedCsvRecord` as convenience methods
 - Rename `errorOnDifferentFieldCount()` to `ignoreDifferentFieldCount()`
 - Rename `isEmpty()` to `isEmptyLine` in `CsvRecord`